Crate loans

Loans pallet

Overview

Loans pallet implement the lending protocol by using a pool-based strategy that aggregates each user’s supplied assets. The interest rate is dynamically determined by the supply and demand.

Re-exports

pub use pallet::*;

pub use weights::WeightInfo;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights

Autogenerated weights for loans

Structs

BorrowSnapshot

Container for borrow balance information

CurveModel

The curve interest rate model

EarnedSnapshot

Container for earned amount information

JumpModel

The jump interest rate model

Market

Market.

OnSlashHook

PostDeposit

PostTransfer

PreDeposit

PreTransfer

RewardMarketState

Enums

InterestRateModel

Parallel interest rate model

MarketState

The current state of a market. For more information, see Market.

Versions

Utility type for managing upgrades/migrations.

Constants

DEFAULT_MAX_EXCHANGE_RATE

DEFAULT_MIN_EXCHANGE_RATE

INCENTIVE_ACCOUNT_PREFIX

REWARD_ACCOUNT_PREFIX

Traits

OnDeposit

Hook to run before depositing into an account.

OnSlash

Hook to run before slashing an account.

OnTransfer

Hook to run before transferring from an account to another.