Enum loans::pallet::Call

pub enum Call<T: Config> {

    add_market {
        asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
        market: Market<<<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance>,
    },
    activate_market {
        asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
    },
    update_rate_model {
        asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
        rate_model: InterestRateModel,
    },
    update_market {
        asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
        collateral_factor: Option<Ratio>,
        liquidation_threshold: Option<Ratio>,
        reserve_factor: Option<Ratio>,
        close_factor: Option<Ratio>,
        liquidate_incentive_reserved_factor: Option<Ratio>,
        liquidate_incentive: Option<Rate>,
        supply_cap: Option<<<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance>,
        borrow_cap: Option<<<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance>,
    },
    force_update_market {
        asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
        market: Market<<<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance>,
    },
    add_reward {
        amount: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance,
    },
    withdraw_missing_reward {
        target_account: <T::Lookup as StaticLookup>::Source,
        amount: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance,
    },
    update_market_reward_speed {
        asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
        supply_reward_per_block: Option<<<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance>,
        borrow_reward_per_block: Option<<<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance>,
    },
    claim_reward {},
    claim_reward_for_market {
        asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
    },
    mint {
        asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
        mint_amount: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance,
    },
    redeem {
        asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
        redeem_amount: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance,
    },
    redeem_all {
        asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
    },
    borrow {
        asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
        borrow_amount: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance,
    },
    repay_borrow {
        asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
        repay_amount: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance,
    },
    repay_borrow_all {
        asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
    },
    deposit_all_collateral {
        asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
    },
    withdraw_all_collateral {
        asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
    },
    liquidate_borrow {
        borrower: T::AccountId,
        liquidation_asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
        repay_amount: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance,
        collateral_asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
    },
    add_reserves {
        payer: <T::Lookup as StaticLookup>::Source,
        asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
        add_amount: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance,
    },
    reduce_reserves {
        receiver: <T::Lookup as StaticLookup>::Source,
        asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
        reduce_amount: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance,
    },
    reduce_incentive_reserves {
        receiver: <T::Lookup as StaticLookup>::Source,
        asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
        redeem_amount: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance,
    },
    // some variants omitted

}

Contains one variant per dispatchable that can be called by an extrinsic.

Variants

add_market

Fields

asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId

market: Market<<<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance>

Stores a new market and its related currency. Returns Err if a currency is not attached to an existent market.

All provided market states must be Pending, otherwise an error will be returned.

If a currency is already attached to a market, then the market will be replaced by the new provided value.

The lend_token id and asset id are bound, the lend_token id of new provided market cannot be duplicated with the existing one, otherwise it will return InvalidLendTokenId.

asset_id: Market related currency

market: The market that is going to be stored

activate_market

Fields

asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId

Activates a market. Returns Err if the market currency does not exist.

If the market is already activated, does nothing.

asset_id: Market related currency

update_rate_model

Fields

asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId

rate_model: InterestRateModel

Updates the rate model of a stored market. Returns Err if the market currency does not exist or the rate model is invalid.

asset_id: Market related currency

rate_model: The new rate model to be updated

update_market

Fields

asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId

collateral_factor: Option<Ratio>

liquidation_threshold: Option<Ratio>

reserve_factor: Option<Ratio>

close_factor: Option<Ratio>

liquidate_incentive_reserved_factor: Option<Ratio>

liquidate_incentive: Option<Rate>

supply_cap: Option<<<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance>

borrow_cap: Option<<<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance>

Updates a stored market. Returns Err if the market currency does not exist.

asset_id: market related currency

collateral_factor: the collateral utilization ratio

reserve_factor: fraction of interest currently set aside for reserves

close_factor: maximum liquidation ratio at one time

liquidate_incentive: liquidation incentive ratio

cap: market capacity

force_update_market

Fields

asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId

market: Market<<<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance>

Force updates a stored market. Returns Err if the market currency does not exist.

asset_id: market related currency

market: the new market parameters

add_reward

Fields

amount: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance

Add reward for the pallet account.

amount: Reward amount added

withdraw_missing_reward

Fields

target_account: <T::Lookup as StaticLookup>::Source

amount: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance

Withdraw reward token from pallet account.

The origin must conform to UpdateOrigin.

target_account: account receive reward token.

amount: Withdraw amount

update_market_reward_speed

Fields

asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId

supply_reward_per_block: Option<<<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance>

borrow_reward_per_block: Option<<<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance>

Updates reward speed for the specified market

The origin must conform to UpdateOrigin.

asset_id: Market related currency

reward_per_block: reward amount per block.

claim_reward

Fields

Claim reward from all market.

claim_reward_for_market

Fields

asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId

Claim reward from the specified market.

asset_id: Market related currency

mint

Fields

asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId

mint_amount: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance

Sender supplies assets into the market and receives internal supplies in exchange.

asset_id: the asset to be deposited.

mint_amount: the amount to be deposited.

redeem

Fields

asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId

redeem_amount: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance

Sender redeems some of internal supplies in exchange for the underlying asset.

asset_id: the asset to be redeemed.

redeem_amount: the amount to be redeemed.

redeem_all

Fields

asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId

Sender redeems all of internal supplies in exchange for the underlying asset.

asset_id: the asset to be redeemed.

borrow

Fields

asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId

borrow_amount: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance

Sender borrows assets from the protocol to their own address.

asset_id: the asset to be borrowed.

borrow_amount: the amount to be borrowed.

repay_borrow

Fields

asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId

repay_amount: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance

Sender repays some of their debts.

asset_id: the asset to be repaid.

repay_amount: the amount to be repaid.

repay_borrow_all

Fields

asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId

Sender repays all of their debts.

asset_id: the asset to be repaid.

deposit_all_collateral

Fields

asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId

withdraw_all_collateral

Fields

asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId

liquidate_borrow

Fields

borrower: T::AccountId

liquidation_asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId

repay_amount: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance

collateral_asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId

The sender liquidates the borrower’s collateral.

borrower: the borrower to be liquidated.

liquidation_asset_id: the assert to be liquidated.

repay_amount: the amount to be repaid borrow.

collateral_asset_id: The collateral to seize from the borrower.

add_reserves

Fields

payer: <T::Lookup as StaticLookup>::Source

asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId

add_amount: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance

Add reserves by transferring from payer.

May only be called from T::ReserveOrigin.

payer: the payer account.

asset_id: the assets to be added.

add_amount: the amount to be added.

reduce_reserves

Fields

receiver: <T::Lookup as StaticLookup>::Source

asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId

reduce_amount: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance

Reduces reserves by transferring to receiver.

May only be called from T::ReserveOrigin.

receiver: the receiver account.

asset_id: the assets to be reduced.

reduce_amount: the amount to be reduced.

reduce_incentive_reserves

Fields

receiver: <T::Lookup as StaticLookup>::Source

asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId

redeem_amount: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance

Sender redeems some of internal supplies in exchange for the underlying asset.

asset_id: the asset to be redeemed.

redeem_amount: the amount to be redeemed.

Implementations

impl<T: Config> Call<T>

pub fn new_call_variant_add_market(
asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
market: Market<<<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance>
) -> Self

Create a call with the variant add_market.

pub fn new_call_variant_activate_market(
asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId
) -> Self

Create a call with the variant activate_market.

pub fn new_call_variant_update_rate_model(
asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
rate_model: InterestRateModel
) -> Self

Create a call with the variant update_rate_model.

pub fn new_call_variant_update_market(
asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
collateral_factor: Option<Ratio>,
liquidation_threshold: Option<Ratio>,
reserve_factor: Option<Ratio>,
close_factor: Option<Ratio>,
liquidate_incentive_reserved_factor: Option<Ratio>,
liquidate_incentive: Option<Rate>,
supply_cap: Option<<<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance>,
borrow_cap: Option<<<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance>
) -> Self

Create a call with the variant update_market.

pub fn new_call_variant_force_update_market(
asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
market: Market<<<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance>
) -> Self

Create a call with the variant force_update_market.

pub fn new_call_variant_add_reward(
amount: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance
) -> Self

Create a call with the variant add_reward.

pub fn new_call_variant_withdraw_missing_reward(
target_account: <T::Lookup as StaticLookup>::Source,
amount: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance
) -> Self

Create a call with the variant withdraw_missing_reward.

pub fn new_call_variant_update_market_reward_speed(
asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
supply_reward_per_block: Option<<<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance>,
borrow_reward_per_block: Option<<<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance>
) -> Self

Create a call with the variant update_market_reward_speed.

pub fn new_call_variant_claim_reward() -> Self

Create a call with the variant claim_reward.

pub fn new_call_variant_claim_reward_for_market(
asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId
) -> Self

Create a call with the variant claim_reward_for_market.

pub fn new_call_variant_mint(
asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
mint_amount: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance
) -> Self

Create a call with the variant mint.

pub fn new_call_variant_redeem(
asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
redeem_amount: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance
) -> Self

Create a call with the variant redeem.

pub fn new_call_variant_redeem_all(
asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId
) -> Self

Create a call with the variant redeem_all.

pub fn new_call_variant_borrow(
asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
borrow_amount: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance
) -> Self

Create a call with the variant borrow.

pub fn new_call_variant_repay_borrow(
asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
repay_amount: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance
) -> Self

Create a call with the variant repay_borrow.

pub fn new_call_variant_repay_borrow_all(
asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId
) -> Self

Create a call with the variant repay_borrow_all.

pub fn new_call_variant_deposit_all_collateral(
asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId
) -> Self

Create a call with the variant deposit_all_collateral.

pub fn new_call_variant_withdraw_all_collateral(
asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId
) -> Self

Create a call with the variant withdraw_all_collateral.

pub fn new_call_variant_liquidate_borrow(
borrower: T::AccountId,
liquidation_asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
repay_amount: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance,
collateral_asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId
) -> Self

Create a call with the variant liquidate_borrow.

pub fn new_call_variant_add_reserves(
payer: <T::Lookup as StaticLookup>::Source,
asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
add_amount: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance
) -> Self

Create a call with the variant add_reserves.

pub fn new_call_variant_reduce_reserves(
receiver: <T::Lookup as StaticLookup>::Source,
asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
reduce_amount: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance
) -> Self

Create a call with the variant reduce_reserves.

pub fn new_call_variant_reduce_incentive_reserves(
receiver: <T::Lookup as StaticLookup>::Source,
asset_id: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId,
redeem_amount: <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance
) -> Self

Create a call with the variant reduce_incentive_reserves.