Module loans::pallet::storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

AccountBorrows

Mapping of account addresses to outstanding borrow balances CurrencyId -> Owner -> BorrowSnapshot

AccountDeposits

Mapping of account addresses to collateral deposit details CollateralType -> Owner -> Collateral Deposits

BorrowIndex

Accumulator of the total earned interest rate since the opening of the market CurrencyId -> u128

BorrowRate

Mapping of borrow rate to currency type

ExchangeRate

The internal exchange rate from the associated lend token to the underlying currency.

LastAccruedInterestTime

The timestamp of the last calculation of accrued interest

Markets

Mapping of underlying currency id to its market

MaxExchangeRate

The maximum allowed exchange rate for a market.

MinExchangeRate

The minimum allowed exchange rate for a market. This is the starting rate when a market is first set up.

RewardAccrued

The incentive reward accrued but not yet transferred to each user.

RewardBorrowSpeed

Mapping of underlying currency id to borrow reward speed

RewardBorrowState

The Reward market borrow state for each market

RewardBorrowerIndex

The incentive reward index for each market for each borrower as of the last time they accrued Reward

RewardSupplierIndex

The incentive reward index for each market for each supplier as of the last time they accrued Reward

RewardSupplySpeed

Mapping of underlying currency id to supply reward speed

RewardSupplyState

The Reward market supply state for each market

StorageVersion

Storage version of the pallet.

SupplyRate

Mapping of supply rate to currency type

TotalBorrows

Total amount of outstanding borrows of the underlying in this market CurrencyId -> Balance

TotalReserves

Total amount of reserves of the underlying held in this market CurrencyId -> Balance

UnderlyingAssetId

Mapping of lend_token id to underlying currency id lend_token id: voucher token id asset id: underlying token id

UtilizationRatio

Borrow utilization ratio