Struct loans::pallet::Pallet

pub struct Pallet<T>(_);

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Implementations

impl<T: Config> Pallet<T>

pub fn reward_account_id() -> T::AccountId

impl<T: Config> Pallet<T>

pub fn accrue_interest(asset_id: <T as Config>::CurrencyId) -> DispatchResult

Accrue interest and update corresponding storage

pub fn get_market_status( asset_id: <T as Config>::CurrencyId ) -> Result<(Rate, Rate, Rate, Ratio, <T as Config>::Balance, <T as Config>::Balance, FixedU128), DispatchError>

pub fn exchange_rate_stored( asset_id: <T as Config>::CurrencyId ) -> Result<Rate, DispatchError>

Update the exchange rate according to the totalCash, totalBorrows and totalSupply. This function does not accrue interest before calculating the exchange rate. exchangeRate = (totalCash + totalBorrows - totalReserves) / totalSupply

impl<T: Config> Pallet<T>

pub fn add_market( origin: OriginFor<T>, asset_id: <T as Config>::CurrencyId, market: Market<<T as Config>::Balance> ) -> DispatchResultWithPostInfo

Creates a new lending market for a given currency. Returns Err if a market already exists for the given currency.

All provided market states must be Pending, otherwise an error will be returned.

The lend_token id specified in the Market struct has to be unique, and cannot be later reused when creating a new market.

asset_id: Currency to enable lending and borrowing for.

market: Configuration of the new lending market

pub fn activate_market( origin: OriginFor<T>, asset_id: <T as Config>::CurrencyId ) -> DispatchResultWithPostInfo

Activates a market. Returns Err if the market does not exist.

If the market is already active, does nothing.

asset_id: Currency to enable lending and borrowing for.

pub fn update_rate_model( origin: OriginFor<T>, asset_id: <T as Config>::CurrencyId, rate_model: InterestRateModel ) -> DispatchResultWithPostInfo

Updates the rate model of a stored market. Returns Err if the market currency does not exist or the rate model is invalid.

asset_id: Market currency

rate_model: The new rate model to set

pub fn update_market( origin: OriginFor<T>, asset_id: <T as Config>::CurrencyId, collateral_factor: Option<Ratio>, liquidation_threshold: Option<Ratio>, reserve_factor: Option<Ratio>, close_factor: Option<Ratio>, liquidate_incentive_reserved_factor: Option<Ratio>, liquidate_incentive: Option<Rate>, supply_cap: Option<<T as Config>::Balance>, borrow_cap: Option<<T as Config>::Balance> ) -> DispatchResultWithPostInfo

Updates a stored market. Returns Err if the market currency does not exist.

asset_id: market related currency

collateral_factor: the collateral utilization ratio

liquidation_threshold: The collateral ratio when a borrower can be liquidated

reserve_factor: fraction of interest set aside for reserves

close_factor: max percentage of debt that can be liquidated in a single transaction

liquidate_incentive_reserved_factor: liquidation share set aside for reserves

liquidate_incentive: liquidation incentive ratio

supply_cap: Upper bound of supplying

borrow_cap: Upper bound of borrowing

pub fn force_update_market( origin: OriginFor<T>, asset_id: <T as Config>::CurrencyId, market: Market<<T as Config>::Balance> ) -> DispatchResultWithPostInfo

Force updates a stored market. Returns Err if the market currency does not exist.

asset_id: market related currency

market: Configuration of the new lending market

pub fn add_reward( origin: OriginFor<T>, amount: <T as Config>::Balance ) -> DispatchResultWithPostInfo

Deposit incentive reward currency into the pallet account.

amount: Reward amount added

pub fn update_market_reward_speed( origin: OriginFor<T>, asset_id: <T as Config>::CurrencyId, supply_reward_per_block: Option<<T as Config>::Balance>, borrow_reward_per_block: Option<<T as Config>::Balance> ) -> DispatchResultWithPostInfo

Updates reward speed for the specified market

The origin must conform to UpdateOrigin.

asset_id: Market related currency

supply_reward_per_block: supply reward amount per block.

borrow_reward_per_block: borrow reward amount per block.

pub fn claim_reward(origin: OriginFor<T>) -> DispatchResultWithPostInfo

Claim incentive rewards for all markets.

pub fn claim_reward_for_market( origin: OriginFor<T>, asset_id: <T as Config>::CurrencyId ) -> DispatchResultWithPostInfo

Claim inceitve reward for the specified market.

asset_id: Market related currency

pub fn mint( origin: OriginFor<T>, asset_id: <T as Config>::CurrencyId, mint_amount: <T as Config>::Balance ) -> DispatchResultWithPostInfo

The caller supplies (lends) assets into the market and receives a corresponding amount of lend tokens, at the current internal exchange rate.

asset_id: the asset to be deposited.

mint_amount: the amount to be deposited.

pub fn redeem( origin: OriginFor<T>, asset_id: <T as Config>::CurrencyId, redeem_amount: <T as Config>::Balance ) -> DispatchResultWithPostInfo

The caller redeems lend tokens for the underlying asset, at the current internal exchange rate.

asset_id: the asset to be redeemed

redeem_amount: the amount to be redeemed, expressed in the underyling currency (asset_id)

pub fn redeem_all( origin: OriginFor<T>, asset_id: <T as Config>::CurrencyId ) -> DispatchResultWithPostInfo

The caller redeems their entire lend token balance in exchange for the underlying asset. Note: this will fail if the account needs some of the collateral for backing open borrows, or if any of the lend tokens are used by other pallets (e.g. used as vault collateral)

asset_id: the asset to be redeemed.

pub fn borrow( origin: OriginFor<T>, asset_id: <T as Config>::CurrencyId, borrow_amount: <T as Config>::Balance ) -> DispatchResultWithPostInfo

The caller borrows borrow_amount of asset_id from the protocol, using their supplied assets as collateral.

asset_id: the asset to be borrowed.

borrow_amount: the amount to be borrowed.

pub fn repay_borrow( origin: OriginFor<T>, asset_id: <T as Config>::CurrencyId, repay_amount: <T as Config>::Balance ) -> DispatchResultWithPostInfo

The caller repays some of their debts.

asset_id: the asset to be repaid.

repay_amount: the amount to be repaid, in the underlying currency (asset_id).

pub fn repay_borrow_all( origin: OriginFor<T>, asset_id: <T as Config>::CurrencyId ) -> DispatchResultWithPostInfo

The caller repays all of their debts.

asset_id: the asset to be repaid.

pub fn deposit_all_collateral( origin: OriginFor<T>, asset_id: <T as Config>::CurrencyId ) -> DispatchResultWithPostInfo

Caller enables their lend token balance as borrow collateral. This operation locks the lend tokens, so they are no longer transferrable. Any incoming lend tokens into the caller’s account (either by direct transfer or minting) are automatically locked as well, such that locking and unlocking borrow collateral is an atomic state (a “collateral toggle”). If any of the caller’s lend token balance is locked elsewhere (for instance, as bridge vault collateral), this operation will fail. If this operation is successful, the caller’s maximum allowed debt increases.

asset_id: the underlying asset denoting the market whose lend tokens are to be enabled as collateral.

pub fn withdraw_all_collateral( origin: OriginFor<T>, asset_id: <T as Config>::CurrencyId ) -> DispatchResultWithPostInfo

Caller disables their lend token balance as borrow collateral. This operation unlocks the lend tokens, so they become transferrable. This operation can only succeed if the caller’s debt is backed by sufficient collateral excluding this currency.

asset_id: the underlying asset denoting the market whose lend tokens are to be disabled as collateral.

pub fn liquidate_borrow( origin: OriginFor<T>, borrower: T::AccountId, liquidation_asset_id: <T as Config>::CurrencyId, repay_amount: <T as Config>::Balance, collateral_asset_id: <T as Config>::CurrencyId ) -> DispatchResultWithPostInfo

The caller liquidates the borrower’s collateral. This extrinsic may need to be called multiple times to completely clear the borrower’s bad debt, because of the close_factor parameter in the market. See the close_factor_may_require_multiple_liquidations_to_clear_bad_debt unit test for an example of this.

borrower: the borrower to be liquidated.

liquidation_asset_id: the underlying asset to be liquidated.

repay_amount: the amount of liquidation_asset_id to be repaid. This parameter can only be as large as the close_factor market parameter allows (close_factor * borrower_debt_in_liquidation_asset).

collateral_asset_id: The underlying currency whose lend tokens to seize from the borrower. Note that the liquidator has to redeem the received lend tokens from the market to convert them to collateral_asset_id.

pub fn add_reserves( origin: OriginFor<T>, payer: <T::Lookup as StaticLookup>::Source, asset_id: <T as Config>::CurrencyId, add_amount: <T as Config>::Balance ) -> DispatchResultWithPostInfo

Add reserves by transferring from payer. TODO: This extrinsic currently does nothing useful. See the TODO comment of the ensure_enough_cash function for more details. Based on that TODO, decide whether this extrinsic should be kept.

May only be called from T::ReserveOrigin.

payer: the payer account.

asset_id: the assets to be added.

add_amount: the amount to be added.

pub fn reduce_reserves( origin: OriginFor<T>, receiver: <T::Lookup as StaticLookup>::Source, asset_id: <T as Config>::CurrencyId, reduce_amount: <T as Config>::Balance ) -> DispatchResultWithPostInfo

Reduces reserves (treasury’s share of accrued interest) by transferring to receiver.

May only be called from T::ReserveOrigin.

receiver: the receiver account.

asset_id: the assets to be reduced.

reduce_amount: the amount to be reduced.

pub fn reduce_incentive_reserves( origin: OriginFor<T>, receiver: <T::Lookup as StaticLookup>::Source, asset_id: <T as Config>::CurrencyId, redeem_amount: <T as Config>::Balance ) -> DispatchResultWithPostInfo

Sender redeems some of internal supplies in exchange for the underlying asset.

asset_id: the asset to be redeemed.

redeem_amount: the amount to be redeemed.

impl<T: Config> Pallet<T>

pub fn last_accrued_interest_time<KArg>(k: KArg) -> Timestampwhere KArg: EncodeLike<<T as Config>::CurrencyId>,

An auto-generated getter for LastAccruedInterestTime.

impl<T: Config> Pallet<T>

pub fn account_borrows<KArg1, KArg2>( k1: KArg1, k2: KArg2 ) -> BorrowSnapshot<<T as Config>::Balance>where KArg1: EncodeLike<<T as Config>::CurrencyId>, KArg2: EncodeLike<T::AccountId>,

An auto-generated getter for AccountBorrows.

impl<T: Config> Pallet<T>

pub fn borrow_index<KArg>(k: KArg) -> Ratewhere KArg: EncodeLike<<T as Config>::CurrencyId>,

An auto-generated getter for BorrowIndex.

impl<T: Config> Pallet<T>

pub fn exchange_rate<KArg>(k: KArg) -> Ratewhere KArg: EncodeLike<<T as Config>::CurrencyId>,

An auto-generated getter for ExchangeRate.

impl<T: Config> Pallet<T>

pub fn borrow_rate<KArg>(k: KArg) -> Ratewhere KArg: EncodeLike<<T as Config>::CurrencyId>,

An auto-generated getter for BorrowRate.

impl<T: Config> Pallet<T>

pub fn supply_rate<KArg>(k: KArg) -> Ratewhere KArg: EncodeLike<<T as Config>::CurrencyId>,

An auto-generated getter for SupplyRate.

impl<T: Config> Pallet<T>

pub fn utilization_ratio<KArg>(k: KArg) -> Ratiowhere KArg: EncodeLike<<T as Config>::CurrencyId>,

An auto-generated getter for UtilizationRatio.

impl<T: Config> Pallet<T>

pub fn reward_supply_speed<KArg>(k: KArg) -> <T as Config>::Balancewhere KArg: EncodeLike<<T as Config>::CurrencyId>,

An auto-generated getter for RewardSupplySpeed.

impl<T: Config> Pallet<T>

pub fn reward_borrow_speed<KArg>(k: KArg) -> <T as Config>::Balancewhere KArg: EncodeLike<<T as Config>::CurrencyId>,

An auto-generated getter for RewardBorrowSpeed.

impl<T: Config> Pallet<T>

pub fn reward_supply_state<KArg>( k: KArg ) -> RewardMarketState<BlockNumberFor<T>, <T as Config>::Balance>where KArg: EncodeLike<<T as Config>::CurrencyId>,

An auto-generated getter for RewardSupplyState.

impl<T: Config> Pallet<T>

pub fn reward_borrow_state<KArg>( k: KArg ) -> RewardMarketState<BlockNumberFor<T>, <T as Config>::Balance>where KArg: EncodeLike<<T as Config>::CurrencyId>,

An auto-generated getter for RewardBorrowState.

impl<T: Config> Pallet<T>

pub fn reward_supplier_index<KArg1, KArg2>( k1: KArg1, k2: KArg2 ) -> <T as Config>::Balancewhere KArg1: EncodeLike<<T as Config>::CurrencyId>, KArg2: EncodeLike<T::AccountId>,

An auto-generated getter for RewardSupplierIndex.

impl<T: Config> Pallet<T>

pub fn reward_borrower_index<KArg1, KArg2>( k1: KArg1, k2: KArg2 ) -> <T as Config>::Balancewhere KArg1: EncodeLike<<T as Config>::CurrencyId>, KArg2: EncodeLike<T::AccountId>,

An auto-generated getter for RewardBorrowerIndex.

impl<T: Config> Pallet<T>

pub fn reward_accrued<KArg>(k: KArg) -> <T as Config>::Balancewhere KArg: EncodeLike<T::AccountId>,

An auto-generated getter for RewardAccrued.

impl<T: Config> Pallet<T>

pub fn max_exchange_rate() -> Rate

An auto-generated getter for MaxExchangeRate.

impl<T: Config> Pallet<T>

pub fn min_exchange_rate() -> Rate

An auto-generated getter for MinExchangeRate.

impl<T: Config> Pallet<T>

pub fn account_deposits( lend_token_id: <T as Config>::CurrencyId, supplier: &T::AccountId ) -> Amount<T>

pub fn total_borrows(asset_id: <T as Config>::CurrencyId) -> Amount<T>

pub fn total_reserves(asset_id: <T as Config>::CurrencyId) -> Amount<T>

pub fn account_id() -> T::AccountId

pub fn get_account_liquidity( account: &T::AccountId ) -> Result<AccountLiquidity<T>, DispatchError>

pub fn get_account_liquidation_threshold_liquidity( account: &T::AccountId ) -> Result<AccountLiquidity<T>, DispatchError>

pub fn current_borrow_balance( who: &T::AccountId, asset_id: <T as Config>::CurrencyId ) -> Result<Amount<T>, DispatchError>

pub fn borrow_balance_from_old_and_new_index( old_index: &FixedU128, new_index: &FixedU128, amount: Amount<T>, rounding: Rounding ) -> Result<Amount<T>, DispatchError>

pub fn do_liquidate_borrow( liquidator: T::AccountId, borrower: T::AccountId, repayment_underlying: &Amount<T>, collateral_asset_id: <T as Config>::CurrencyId ) -> DispatchResult

Note:

liquidation_asset_id is borrower’s debt asset.

collateral_asset_id is borrower’s collateral asset.

repay_amount is amount of liquidation_asset_id

The liquidator will repay a certain amount of liquidation_asset_id from own account for borrower. Then the protocol will reduce borrower’s debt and liquidator will receive collateral_asset_id (as voucher amount) from borrower.

pub fn lock_if_account_deposited( account_id: &T::AccountId, lend_tokens: &Amount<T> ) -> DispatchResult

pub fn total_supply( asset_id: <T as Config>::CurrencyId ) -> Result<Amount<T>, DispatchError>

Total issuance of lending tokens (lend_tokens), given the underlying

pub fn free_lend_tokens( asset_id: <T as Config>::CurrencyId, account_id: &T::AccountId ) -> Result<Amount<T>, DispatchError>

Free lending tokens (lend_tokens) of an account, given the underlying

pub fn reserved_lend_tokens( asset_id: <T as Config>::CurrencyId, account_id: &T::AccountId ) -> Result<Amount<T>, DispatchError>

Reserved lending tokens (lend_tokens) of an account, given the underlying

pub fn get_asset_value(asset: &Amount<T>) -> Result<Amount<T>, DispatchError>

pub fn market( asset_id: <T as Config>::CurrencyId ) -> Result<Market<<T as Config>::Balance>, DispatchError>

pub fn lend_token_id( asset_id: <T as Config>::CurrencyId ) -> Result<<T as Config>::CurrencyId, DispatchError>

pub fn incentive_reward_account_id() -> T::AccountId