Type Definition loans::pallet::RewardBorrowerIndex

pub type RewardBorrowerIndex<T: Config> = StorageDoubleMap<_GeneratedPrefixForStorageRewardBorrowerIndex<T>, Blake2_128Concat, <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::AssetId, Blake2_128Concat, T::AccountId, <<T as Config>::Assets as Inspect<<T as Config>::AccountId>>::Balance, ValueQuery>;

The Reward index for each market for each borrower as of the last time they accrued Reward